backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
requests
google-genai